pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "neural_astar_repro"
version = "0.2.0"
description = "Neural A* (ICML 2021) — standalone reproduction without Hydra/Lightning"
# Paper code is fine on 3.10–3.13; 3.14 may still hit unrelated wheel/build issues — use a stable CPython.
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy>=1.22",
    "segmentation-models-pytorch>=0.3.3",
    "pqdict>=1.2.0",
    "tqdm>=4.66",
    "pillow>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
viz = ["moviepy>=1.0.3"]
